refactor: unify credential-source parsing into one neutral module
The operator credential-source vocabulary (env:/file:/command:) was parsed in
git_hosts/credentials.py, which then reached across a layer to resolve secrets
via the private inner.credential_proxy._resolve_secret. Consolidate parse,
validate, and resolve into a neutral top-level omnigent.credential_sources
module that both the server-side host config and the in-sandbox credential
proxy import, and drop the cross-layer private import. The accepted/rejected
source-string set is byte-identical.
